hw6 /requirements.txt
gymnasium[box2d, mujoco]==0.29.1
tensorboard==2.10.0
tensorboardX==2.5.1
matplotlib>=3.8,<3.9
ipython==7.34.0
moviepy==1.0.3
pyvirtualdisplay==3.0
torch>=2.2.0
opencv-python==4.6.0.66
ipdb==0.13.9
swig==4.0.2
tqdm==4.66.1
pyyaml==6.0.1
numpy<2
